requirements.txt
# Core framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1

# LLM Providers
openai==1.59.5
anthropic==0.42.0

# Vector Store
chromadb==0.5.23

# Document Processing
PyMuPDF==1.25.3
python-docx==1.1.2

# Embeddings (used by ChromaDB)
sentence-transformers==3.3.1

# Utilities
python-dotenv==1.0.1
httpx==0.28.1

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0
